Improved initialisation
relp_initialise now fails in a sensible way if the client does not
start with an open command or if a few other conditions are not
met. In the case of failure the whole connection is closed and we
wait for a new one to be made (as per the specification re. errors)