Improved the RelpClient interface
I've fiddled the model around slightly so that it is non-blocking
unless its buffer has filled up. Returning the unacked frames on
close is a bit dodgey, but does allow some way to recover them in
case of problems. I'm still not all that happy with RelpClient so
it is liable to be changed again