outputs/riemann: make event tagging work again
As tags on outputs are now deprecated, we pass the tags set at input or
filter stage to Riemann events.

For the record, the idea is to pass along Logstash's events tags to the
Riemann events created, so that filtering *in Riemann* can be done based
on the tags set by Logstash.