Bug // Compress Spooler will no longer drop messages - issue#110
When the last message is ready to be spooled, compressed spooler will
drop the message it's currently processing and push out the spool. If
your spool size is 10, this means on the 11th message it will send that
spool.

This changes the logic to add the message to the spool, and then
immediately send it if it's the right size.