Start Event Field and End Event Timestamp
I had a usecase where i needed to add a special field from the start event to the elapsed event. 
And I needed to add the end event timestamp to the elapsed event aswell because the time difference was sometimes less than a few milliseconds which does not show if the elapsed time is shown in seconds.